requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
tensorflow
google-generativeai